fix: redundant serde functions and enable bin serializtion
